+++ libs/feature-orden-compra/src/lib/feature-orden-compra/feature-orden-compra.ts
@@ -1,0 +1,9 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'lib-feature-compras-pagos',
+  imports: [],
+  templateUrl: './feature-compras-pagos.html',
+  styleUrl: './feature-compras-pagos.css'
+})
+export class FeatureComprasPagos {}